+++ csv-processor/src/main/java/com/smdev/csv/CsvFileReader.java
@@ -8,9 +8,9 @@ import java.util.List;
 import com.opencsv.CSVReader;
 import com.smdev.exc.ApplicationException;
 import com.smdev.exc.ReadFileException;
+import com.smdev.file.SMFileReader;
 import com.smdev.model.DataCellType;
 import com.smdev.model.Data;
-import com.smdev.processor.SMFileReader;
 
 /**
  * Implementation of a CSV file reader.
