+++ data-model/src/main/java/com/smdev/file/SMFileReader.java
@@ -1,4 +1,4 @@
-package com.smdev.processor;
+package com.smdev.file;
 
 import java.io.File;
