+++ ms-processor/src/main/java/com/smdev/ms/excel/ExcelFileWriter.java
@@ -9,8 +9,8 @@ import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 
 import com.smdev.exc.ApplicationException;
 import com.smdev.exc.WriteFileExaception;
+import com.smdev.file.SMFileWriter;
 import com.smdev.model.Data;
-import com.smdev.processor.SMFileWriter;
 import com.smdev.util.ResourceUtils;
 
 public class ExcelFileWriter implements SMFileWriter<ExcelProps> {
